Use regex to validate a mcma field.
It's easier to set up the correct answers in Constants first.
We check by trying to remove correct answers a copy of the player's
answer.
If the original answer is correct, then there should be no word
character ([0-9A-Za-z]) left in the answer and regex should return
`None`.
Otherwise it returns the prompt.